major restructuring and cleanup
- many inner classes moved to their own file
- packages added to help organize classes
- static methods moved to more specific or more intuitive places
- Codec.Set renamed Codec.Field
- Codec.TYPE enum removed
- getSingleton methods replaced with public INSTANCE constants
