+++ src/main/java/com/addthis/codec/json/CodecExceptionLineNumber.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.addthis.codec;
+package com.addthis.codec.json;
 
 import javax.annotation.Nonnull;
