+++ src/main/java/com/addthis/codec/util/CodableObjectWriter.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.addthis.codec;
+package com.addthis.codec.util;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -19,6 +19,8 @@ import java.io.OutputStream;
 import java.util.zip.GZIPOutputStream;
 
 import com.addthis.basis.util.Bytes;
+
+import com.addthis.codec.Codec;
 
 public class CodableObjectWriter<T> {
