+++ src/main/java/com/addthis/codec/validation/ValidationException.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.addthis.codec.validation;
+
+import com.addthis.codec.reflection.PolicyException;
+
+@SuppressWarnings("serial")
+public class ValidationException extends PolicyException {
+
+    private String field;
+
+    public ValidationException(String msg, String field) {
+        super(msg);
+        this.field = field;
+    }
+
+    public String getField() {
+        return field;
+    }
+}
